+++ src/models/RetconTransformedImage.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace mmikkel\retcon\models;
+
+use craft\base\Model;
+
+class RetconTransformedImage extends Model
+{
+
+    /** @var string */
+    public $url;
+
+    /** @var int */
+    public $width;
+
+    /** @var int */
+    public $height;
+
+}