+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 ### Added
-- Added a new `$allowFilenameAsAltText` parameter to the `retconAutoAlt` filter (default `true`; will be default `false` in Retcon 3.0)  
+- Added a new `$allowFilenameAsAltText` parameter to the `retconAutoAlt` filter (default `true`; will be default `false` in Retcon 3.0).  
+- Added `RetconTransformedImage` model.  
 ### Improved
 - Improved support for CKEditor.
 ### Changed
@@ -13,6 +14,7 @@
 - Fixed an issue where the `retconUnwrap` filter could unwrap root nodes.
 - Fixed an issue where the `autoAlt` filter was unable to retrieve alt value from the source asset.
 - Fixed an issue where the `retconAutoAlt` filter could use base64 `src` placeholders could be used for the alt value
+- Fixed a PHP exception that would occur if using `retconTransform` or `retconSrcset` with Imager X named transforms, if the named transform consisted of multiple transforms. Retcon will now use the first transform whenever Imager returns multiple transforms.  
 
 ## 2.6.1 - 2023-02-17
 ### Fixed